// src/services/api/generated.ts
export interface TodoItem {
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)$
   */
  id?: number | string | null;
  title?: string;
  description?: string;
  isCompleted?: boolean;
  category?: string;
  color?: string;
  createdAt?: string;
}